# back/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Oasix property management system API server.",
        "title": "Oasix API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/media/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Keycloak": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get media file information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media"
                ],
                "summary": "Get media info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Get media parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.DefaultRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.MediaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/media/{id}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Keycloak": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Download a file from S3 storage",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "media"
                ],
                "summary": "Download file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/media/{id}/raw": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Keycloak": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Raw a file from S3 storage",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "media"
                ],
                "summary": "Raw file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_handlers.MediaResponse": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type_id": {
                    "type": "string"
                },
                "type_name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "parier-server_internal_models.DefaultRequest": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/parier-server_internal_models.User"
                }
            }
        },
        "parier-server_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "parier-server_internal_models.User": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "email": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "realm": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "session_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2Keycloak": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "http://localhost:18080/realms/parier/protocol/openid-connect/auth",
            "scopes": {
                "read_access": "Grants read access",
                "write_access": "Grants write access"
            }
        }
    }
}